+++ sdk/appcenter-push/src/main/java/com/microsoft/appcenter/push/Push.java
@@ -18,6 +18,10 @@ import android.support.annotation.VisibleForTesting;
 import android.support.annotation.WorkerThread;
 import android.util.Log;
 
+import com.google.android.gms.tasks.OnFailureListener;
+import com.google.android.gms.tasks.OnSuccessListener;
+import com.google.firebase.iid.FirebaseInstanceId;
+import com.google.firebase.iid.InstanceIdResult;
 import com.microsoft.appcenter.AbstractAppCenterService;
 import com.microsoft.appcenter.Flags;
 import com.microsoft.appcenter.UserInformation;
@@ -266,8 +270,9 @@ public class Push extends AbstractAppCenterService {
      *
      * @param pushToken the push token value.
      */
-    synchronized void onTokenRefresh(final String pushToken) {
-        if (pushToken != null) {
+    @SuppressWarnings("WeakerAccess") /* protected so that Xamarin can use it. */
+    protected synchronized void onTokenRefresh(final String pushToken) {
+        if (pushToken != null && !pushToken.equals(mLatestPushToken)) {
             AppCenterLog.debug(LOG_TAG, "Push token refreshed: " + pushToken);
             mLatestPushToken = pushToken;
             post(new Runnable() {
@@ -481,11 +486,12 @@ public class Push extends AbstractAppCenterService {
 
         /* Update enable state of the firebase service. */
         FirebaseUtils.setFirebaseMessagingServiceEnabled(mContext, FirebaseUtils.isFirebaseAvailable());
+        FirebaseInstanceId firebaseInstanceId;
         try {
 
             /* Try to get token through firebase. */
-            onTokenRefresh(FirebaseUtils.getToken());
             AppCenterLog.info(LOG_TAG, "Firebase SDK is available, using Firebase SDK registration.");
+            firebaseInstanceId = FirebaseUtils.getFirebaseInstanceId();
         } catch (FirebaseUtils.FirebaseUnavailableException e) {
             AppCenterLog.warn(LOG_TAG, "Firebase SDK is not available, using built in registration. " +
                     "For all the Android developers using App Center, there is a change coming where Firebase SDK is required " +
@@ -494,7 +500,35 @@ public class Push extends AbstractAppCenterService {
                     "Please follow the migration guide at https://aka.ms/acfba.\n" +
                     "Cause: " + e.getMessage());
             registerPushTokenWithoutFirebase();
+            return;
         }
+
+        /* Use the current API. */
+        try {
+            firebaseInstanceId.getInstanceId().addOnSuccessListener(new OnSuccessListener<InstanceIdResult>() {
+
+                @Override
+                public void onSuccess(InstanceIdResult instanceIdResult) {
+                    onTokenRefresh(instanceIdResult.getToken());
+                }
+            }).addOnFailureListener(new OnFailureListener() {
+
+                @Override
+                public void onFailure(@NonNull Exception e) {
+                    AppCenterLog.error(LOG_TAG, "Failed to register push.", e);
+                }
+            });
+        } catch (NoSuchMethodError e) {
+            onTokenRefresh(getToken(firebaseInstanceId));
+        }
+    }
+
+    @SuppressWarnings("deprecation")
+    private String getToken(FirebaseInstanceId firebaseInstanceId) {
+
+        /* On Xamarin, the stable Nuget still uses a version of Firebase having the old API only. */
+        AppCenterLog.debug(LOG_TAG, "Using old Firebase methods.");
+        return firebaseInstanceId.getToken();
     }
 
     /**